# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    AgentActivityState:
      type: string
      enum:
      - IDLE
      - BUSY
      - WAITING
      - ENDED
    ApiError:
      type: object
      description: "The error envelope: a message, plus a reason on a 409."
      properties:
        message:
          type: string
          description: "What went wrong, in words. Always present."
        reason:
          $ref: "#/components/schemas/Reason"
          type: string
          description: "Which refusal this is. Present on the 409s the release and\
            \ integrate flows raise, absent on every other error. A value a client\
            \ does not recognise is shown as the message verbatim."
        conflicts:
          type: array
          items:
            type: string
          description: The conflicted paths. Present only for CONFLICT and MERGE_CONFLICT.
...
    CreateWorkspaceRequest:
      type: object
      required:
      - repositoryId
      - id
      properties:
        repositoryId:
          type: string
          pattern: \S
        id:
          type: string
          pattern: \S
        parent:
          type: string
        branch:
          type: string
        preamble:
          type: string
        adoptExisting:
          type: boolean
    DiscardWorkspaceRequest:
      type: object
      properties:
        result:
          type: string
    Entry:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceDto"
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    IntegrateRequest:
      type: object
      required:
      - summary
      properties:
        summary:
          type: string
          maxLength: 100
          pattern: \S
...
    MergeWorkspaceRequest:
      type: object
      properties:
        target:
          type: string
...
    Reason:
      type: string
      enum:
      - CONFLICT
      - MERGE_CONFLICT
      - NOT_FAST_FORWARD
      - ALREADY_INTEGRATED
      - PUSH_REJECTED
      - VERSION_ALREADY_RELEASED
      - RELEASE_REQUIRED
...
    ReleaseRequest:
      type: object
      required:
      - summary
      properties:
        summary:
          type: string
          maxLength: 100
          pattern: \S
    Response:
      type: object
      properties:
        version:
          type: string
        commitSha:
          type: string
        branch:
          type: string
...
    Response10:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceDto"
        technicalProcessId:
          type: string
    Response11:
      type: object
      properties:
        success:
          type: boolean
    Response12:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceDto"
    Response13:
      type: object
      properties:
        technicalProcessId:
          type: string
...
    Response5:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceDto"
        technicalProcessId:
          type: string
    Response6:
      type: object
      properties:
        commitHash:
          type: string
        hasConflicts:
          type: boolean
        output:
          type: string
    Response7:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
    Response8:
      type: object
      properties:
        commitSha:
          type: string
        branch:
          type: string
        targetBranch:
          type: string
    Response9:
      type: object
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceDto"
...
    WorkspaceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workspaceId:
          type: string
        parent:
          type: string
        branch:
          type: string
        ahead:
          type: integer
          format: int32
        behind:
          type: integer
          format: int32
        conflictsWithParent:
          type: boolean
        status:
          $ref: "#/components/schemas/WorkspaceStatus"
        runtimeStatus:
          $ref: "#/components/schemas/WorkspaceRuntimeStatus"
        runtimeError:
          type: string
        clean:
          type: boolean
        agentActivity:
          $ref: "#/components/schemas/AgentActivityState"
        preamble:
          type: string
        result:
          type: string
        resolvedAt:
          $ref: "#/components/schemas/Instant"
        daemonConnectedAt:
          $ref: "#/components/schemas/Instant"
        daemonVersion:
          type: string
        daemonBuildTime:
          $ref: "#/components/schemas/Instant"
        daemonOutdated:
          type: boolean
...
    WorkspaceRuntimeStatus:
      type: string
      enum:
      - RUNNING
      - STOPPED
      - PROVISIONING
      - FAILED
    WorkspaceStatus:
      type: string
      enum:
      - ACTIVE
      - INTEGRATED
      - ABANDONED
paths:
...
  /workspaces/api/workspaces:
    get:
      parameters:
      - name: repositoryId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response7"
      summary: List
      tags:
      - Workspace Controller
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response12"
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: The workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response9"
        "404":
          description: No such ACTIVE workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Get
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}/active-process:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response13"
      summary: Active Process
      tags:
      - Workspace Controller
...
  /workspaces/api/workspaces/{id}/delete-container:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDto"
      summary: Delete Container
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}/discard:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscardWorkspaceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response11"
        "400":
          description: Bad Request
      summary: Discard
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}/ensure-container:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response10"
      summary: Ensure Container
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}/integrate:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrateRequest"
        required: true
      responses:
        "200":
          description: Merged into the parent branch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response8"
        "400":
          description: "No summary, an oversized one, or a workspace with no branch\
            \ to integrate."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: "Nothing landed and the target is unchanged — including when\
            \ the parent is the default branch, which only /release may write. `reason`\
            \ says which refusal."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Integrate
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}/merge:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeWorkspaceRequest"
        required: true
      responses:
        "200":
          description: Merged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response6"
        "409":
          description: The target is the repository's default branch. Use /integrate
            or /release.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Bad Request
      summary: Merge
      tags:
      - Workspace Controller
...
  /workspaces/api/workspaces/{id}/recreate-container:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response5"
      summary: Recreate Container
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}/release:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseRequest"
        required: true
      responses:
        "200":
          description: Released; the version and the merge commit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: "No summary, an oversized one, or a workspace with no branch\
            \ to release."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Nothing was released and the default branch is unchanged. `reason`
            says which refusal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
      summary: Release
      tags:
      - Workspace Controller
  /workspaces/api/workspaces/{id}/stop-container:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDto"
      summary: Stop Container
      tags:
      - Workspace Controller
info:
  title: qits-workspaces API
  version: 1.0.0-SNAPSHOT